feat: tidying code, tweaking menu functionality
