+++ packages/datum-web/apps/operator/src/app/(protected)/dashboard/page.tsx
@@ -1,6 +1,11 @@
+import { Metadata } from 'next'
 import React from 'react'
 
 import DashboardPage from '@/components/pages/protected/workspace/dashboard/dashboard'
+
+export const metadata: Metadata = {
+  title: 'Dashboard',
+}
 
 const DashboardLanding = () => {
   return <DashboardPage />
