+++ packages/datum-web/apps/operator/src/app/(protected)/workspace/page.tsx
@@ -1,6 +1,11 @@
+import { Metadata } from 'next'
 import React from 'react'
 
 import WorkspacePage from '@/components/pages/protected/workspace/workspace-page'
+
+export const metadata: Metadata = {
+  title: 'My Workspaces',
+}
 
 const WorkspaceLanding: React.FC = () => {
   return <WorkspacePage />
