+++ packages/datum-web/apps/operator/src/components/pages/protected/workspace/marketing/lists/lists-table.tsx
@@ -1,25 +1,26 @@
 'use client'
 
 import Link from 'next/link'
-import { useState, useEffect } from 'react'
+import { useEffect, useState } from 'react'
 
 import { getPathWithParams } from '@repo/common/routes'
 import { OPERATOR_APP_ROUTES } from '@repo/constants'
+import type { Datum } from '@repo/types'
 import { Checkbox } from '@repo/ui/checkbox'
 import {
   ColumnDef,
-  FilterFn,
-  SortingFn,
+  ColumnFiltersState,
   DataTable,
   DataTableColumnHeader,
-  sortingFns,
-  rankItem,
-  compareItems,
-  ColumnFiltersState,
   Row,
 } from '@repo/ui/data-table'
 import { Tag } from '@repo/ui/tag'
-import type { Datum } from '@repo/types'
+
+import {
+  booleanFilter,
+  fuzzyFilter,
+  fuzzySort,
+} from '@/utils/filters/functions'
 
 import ListsTableDropdown from './lists-table-dropdown'
 import { tableStyles } from './page.styles'
@@ -35,40 +36,9 @@ type ListsTableProps = {
 
 const { header, checkboxContainer, link } = tableStyles()
 
-const fuzzyFilter: FilterFn<any> = (row, columnId, value, addMeta) => {
-  if (!value || value === '') return true
-
-  const cellValue = row.getValue(columnId)
-
-  if (!cellValue) return false
-
-  const itemRank = rankItem(cellValue, value)
-
-  addMeta({
-    itemRank,
-  })
-
-  return itemRank.passed
-}
-
-const emptyFilter: FilterFn<any> = (row, columnId, value) => Boolean(value)
-
-const fuzzySort: SortingFn<any> = (rowA, rowB, columnId) => {
-  let dir = 0
-
-  if (rowA.columnFiltersMeta[columnId]) {
-    dir = compareItems(
-      rowA.columnFiltersMeta[columnId]?.itemRank!,
-      rowB.columnFiltersMeta[columnId]?.itemRank!,
-    )
-  }
-
-  return dir === 0 ? sortingFns.alphanumeric(rowA, rowB, columnId) : dir
-}
-
 const filterFns = {
   fuzzy: fuzzyFilter,
-  empty: emptyFilter,
+  empty: booleanFilter,
 }
 
 export const LIST_COLUMNS: ColumnDef<Datum.List>[] = [
