+++ packages/datum-web/apps/operator/src/utils/filters/schemas.tsx
@@ -1,9 +1,9 @@
 import {
   Building,
   ChartLine,
-  Clock11,
   Clock4,
   Clock8,
+  Clock11,
   Coins,
   Cpu,
   Factory,
@@ -21,8 +21,8 @@ import {
   User,
   UserCheck,
   UserRoundCog,
-  Users,
   UserSquare2,
+  Users,
   WalletCards,
 } from 'lucide-react'
